BAEL-1758: removing helper class and adding static functions to the Kotlin Enum example
