+++ core-kotlin/src/main/kotlin/com/baeldung/enums/CardType.kt
@@ -29,5 +29,20 @@ enum class CardType(val color: String) : ICardLimit {
         }
     };
 
+    companion object {
+        fun getCardTypeByColor(color: String): CardType? {
+            for (cardType in CardType.values()) {
+                if (cardType.color.equals(color)) {
+                    return cardType;
+                }
+            }
+            return null
+        }
+
+        fun getCardTypeByName(name: String): CardType {
+            return CardType.valueOf(name.toUpperCase())
+        }
+    }
+
     abstract fun calculateCashbackPercent(): Float
 }
